pyproject.toml
[project]
name = "clipbd-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.66.0",
    "bs4>=0.0.2",
    "copykitten>=1.2.2",
    "dotenv>=0.9.9",
    "firecrawl-py>=2.16.5",
    "hf-transfer>=0.1.9",
    "html-to-markdown[lxml]>=1.9.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "lxml>=6.0.0",
    "openai>=1.99.9",
    "protobuf>=6.32.0",
    "pyinstaller>=6.15.0",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "torch>=2.8.0",
    "transformers>=4.55.2",
    "youtube-transcript-api>=1.2.2",
    "yt-dlp>=2025.9.5",
]


[tool.ruff]
extend = "~/.config/ruff/ruff.toml"
extend-exclude = [
   "**/*.ipynb",
]
lint.extend-ignore = ["PLR2004"]